+++ CHANGES.md
@@ -4,3 +4,4 @@ Next release
 - `Py.import` is an alias for `Py.Import.import_module`.
 - Use `*_opt` naming convention for the functions that return an option
   instead of an exception: `Py.import_opt`, `Py.Object.find_opt`,...
+- of_seq/to_seq converters